bring in work to test on CI

+++ include/ccmath/math/fmanip/scalbn.hpp
@@ -48,10 +48,10 @@ namespace ccm
 	constexpr T scalbn(T num, long exp) noexcept
 	{
 		#if defined(__GNUC__) && (__GNUC__ > 6 || (__GNUC__ == 6 && __GNUC_MINOR__ >= 1)) && !defined(__clang__)
-		if constexpr (std::is_same_v<T, float>) { return __builtin_scalbnf(num, exp); }
-		if constexpr (std::is_same_v<T, double>) { return __builtin_scalbn(num, exp); }
-		if constexpr (std::is_same_v<T, long double>) { return __builtin_scalbnl(num, exp); }
-		return static_cast<T>(__builtin_scalbnl(num, exp));
+		if constexpr (std::is_same_v<T, float>) { return __builtin_scalbnf(num, static_cast<int>(exp)); }
+		if constexpr (std::is_same_v<T, double>) { return __builtin_scalbn(num, static_cast<int>(exp)); }
+		if constexpr (std::is_same_v<T, long double>) { return __builtin_scalbnl(num, static_cast<int>(exp)); }
+		return static_cast<T>(__builtin_scalbnl(num, static_cast<int>(exp)));
 		#else
 		if constexpr (std::is_same_v<T, float>) { return internal::scalbn_float(num, exp); }
 		if constexpr (std::is_same_v<T, double>) { return internal::scalbn_double(num, exp); }
